+++ changelog.md
@@ -1,6 +1,7 @@
 # 0.4.0
 
 * Added config to hide other players paintings
+* Added permissions for server ops to delete other paintings
 
 # 0.3.1